wip: project loading indicator
